+++ CHANGELOG.rst
@@ -1,6 +1,18 @@
 Changelog
 *********
 
+0.30.0 (2022-02-05)
+-------------------
+
+* Update :command:`fuc-find` command to allow users to control whether to use recursive retrieving.
+* Add new command :command:`ngs-trim`.
+* Add new command :command:`ngs-quant`.
+* Add new submodule ``pykallisto``.
+* Update :meth:`pycov.CovFrame.from_bam` method to use filename as sample name when the SM tag is missing.
+* Add new method :meth:`pyvcf.row_phased`. From now on, it's used to get the ``pyvcf.VcfFrame.phased`` property.
+* Add new method :meth:`pyvcf.split` and :command:`vcf-split` command for splitting VCF by individual.
+* Update :meth:`pyvcf.merge` method, :meth:`pyvcf.VcfFrame.merge` method, and :command:`vcf-merge` command to automatically handle the 'chr' string.
+
 0.29.0 (2021-12-19)
 -------------------